a11y atspi: Forward new QAccessibleAnnouncementEvent to AT-SPI
Forward the newly added QAccessibleAnnouncementEvent
to the AT-SPI level as a corresponding AT-SPI Announcement
event.

The AtspiLive enum to specify the priority/politeness
level was added in at-spi2-core/libatspi commit [1] which is
contained in libatspi versions >= 2.50, so define
the relevant values ATSPI_LIVE_ASSERTIVE and ATSPI_LIVE_POLITE
manually for older libatspi versions.

(Note that the mention of binary compatibility in above-mentioned
at-spi2-core commit only applies for ATK, not AT-SPI, so is not
relevant here.)

This makes announcement from the example app attached to QTBUG-75003
work when using the Orca screen reader on Linux.

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/commit/32c9420ec82f07da8033c675a2bbf25e535f0509

Task-number: QTBUG-75003
Change-Id: I752aadca2abdda58b3869e17e74fcd9d7572f915
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>