a11y: Add new QAccessibleAnnouncementEvent
Add a new QAccessibleAnnouncementEvent that
can be used to request the announcement of
a message by assistive technologies
(most prominently screen readers).

The priority/urgency of the message can
be specified using one of the values of the
newly added AnnouncementPriority enum.

As suggested by the maintainer of the Orca screen reader
in a similar merge request for GTK [1], add some
clarification that assertive priority should only be
used for things that are actually worth interrupting
the user's current task immediately rather than waiting
for the next graceful opportunity like the end of the
sentence currently spoken by the screen reader.

All of AT-SPI on Linux, UIA on Windows and NSAccessibility
on macOS have corresponding API:

* Linux/AT-SPI: Announcement event (initially added in
  commit [2], possibility to set priority/politeness level
  added in commit [3])

* Windows/UIA: UiaRaiseNotificationEvent [4]

* macOS/NSAccessibility: announcementRequested
  function [5]

Bridging to the corresponding platform API in the
platform backends is left for separate upcoming
commits.

[1] https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/6461#note_1993454
[2] https://gitlab.gnome.org/GNOME/at-spi2-core/-/commit/26835da29918be2c0da6c08fb70d893de22891d2
[3] https://gitlab.gnome.org/GNOME/at-spi2-core/-/commit/32c9420ec82f07da8033c675a2bbf25e535f0509
[4] https://learn.microsoft.com/en-us/windows/win32/api/uiautomationcoreapi/nf-uiautomationcoreapi-uiaraisenotificationevent
[5] https://developer.apple.com/documentation/appkit/nsaccessibility/notification/1530633-announcementrequested

[ChangeLog][QtGui][QAccessibleAnnouncementEvent] Added new
QAccessibleAnnouncementEvent that can be used to request
the announcement of a message by assistive technologies.

Task-number: QTBUG-75003
Change-Id: Ief60ae7c5a9c8bf8404ad9b2749e253acc5d77c5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>