a11y atspi: Include atspi.h instead of atspi-constants.h
The atspi.h header includes atspi-constants.h and other
libatspi headers.

Include the former here instead of manually including the latter,
in order to also implicitly include atspi-version.h added in
libatspi 2.50 commit [1]:

    commit 927344a34cd5bf81fc64da4968241735ecb4f03b
    Author: Mike Gorse <mgorse@suse.com>
    Date:   Thu Jul 20 15:02:03 2023 -0500

        Add atspi_get_version() to return the runtime version of the AT-SPI library

        Also adds ATSPI_MAJOR_VERSION, ATSPI_MINOR_VERSION, and ATSPI_MICRO_VERSION.

        Fixes #129

This is in order to be able to use the new version defines for version-
specific code in upcoming commits.

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/commit/927344a34cd5bf81fc64da4968241735ecb4f03b

Task-number: QTBUG-75003
Change-Id: If2d21c7dcff291c93fed2aa0d7c8f7acd7fae77d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>