QAbstractItemView: don't calc update rect when there are to many indexes
With 8de62d34321cd827e60b0a7b6e7434070de301ae we try to avoid a complete
repaint by checking if the changed indexes are inside the visible rect.
This works fine when the amount of indexes to check is small but fails
otherwise. Therefore add a configurable, upper bound for this check.

Fixes: QTBUG-124173
Change-Id: I1aedd2cbe6e2d0090a640c8a1cf1cdb69f5b002a
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>