Enable resetting a sort model by descending order
By this change, sort(any_negative_number) will stay unchanged, and it
unsorts the model so that the order will be the same as the source
model.
But when the second argument of the sort is set to DescendingOrder
deliberately, the result will be the same as the source model, but in
reverse order.

[ChangeLog][QTCore][QSortFilterProxyModel] QSortFilterProxyModel.sort
sorts the proxy by the source model index descending if the sort
column is less than zero and the order is descending.

Fixes: QTBUG-124863
Change-Id: I73f9d3b500a929efa969cf24ef3d8fe758167645
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>